backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "daphne>=4.2.0",
    "django>=5.2.3",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "psycopg2-binary>=2.9.10",
    "whitenoise>=6.9.0",
]
